+++ Graphs/GraphEdge.php
@@ -1,0 +1,8 @@
+<?php
+
+class GraphEdge
+{
+    public string $start;
+    public string $end;
+    public int $weight;
+}
