+++ tests/DataStructures/ReverseLinkedListTest.php
@@ -3,11 +3,12 @@
 namespace DataStructures;
 
 require_once __DIR__ . '/../../vendor/autoload.php';
-require_once __DIR__ . '/../../DataStructures/LinkedListItem.php';
-require_once __DIR__ . '/../../DataStructures/ReverseLinkedList.php';
+require_once __DIR__ . '/../../DataStructures/ReverseLinkedList/LinkedListItem.php';
+require_once __DIR__ . '/../../DataStructures/ReverseLinkedList/ReverseLinkedList.php';
 
-use LinkedListItem;
-use ReverseLinkedList;
+
+use DataStructures\ReverseLinkedList\LinkedListItem;
+use DataStructures\ReverseLinkedList\ReverseLinkedList;
 use PHPUnit\Framework\TestCase;
 
 class ReverseLinkedListTest extends TestCase
