+++ tests/Graphs/BellmanFordTest.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once __DIR__ . '/../../vendor/autoload.php';
+require_once __DIR__ . '/../../Graphs/GraphEdge.php';
 require_once __DIR__ . '/../../Graphs/BellmanFord.php';
 
 use PHPUnit\Framework\TestCase;
@@ -24,7 +25,7 @@ class BellmanFordTest extends TestCase
         #prepare array of edges listed by edge start to simplify Bellman-Ford updating weights of other edges
         $edges = [];
         foreach ($edgesRaw as $edgeRaw) {
-            $edge = new Edge();
+            $edge = new GraphEdge();
             $edge->start = $edgeRaw[0];
             $edge->end = $edgeRaw[2];
             $edge->weight = $edgeRaw[1];
